move footer files to folder
