+++ app/client/components/page.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import palette from "../colours";
-import { Nav } from "./nav";
 import { minWidth } from "../styles/breakpoints";
+import { Nav } from "./nav";
 
 // Standard width, centered container
 export const PageContainer: React.SFC<{}> = ({ children }) => {
